Fixed ver32 encoding of schema dictionary versions
Version strings for schema dictionaries where one of the major, minor,
or update values is 10 or greater are incorrectly encoded into the ver32
format as described in DSP0240. For example, the version string
'v1_10_0' is encoded with a big-endian representation of 0xF1FAF000
instead of 0xF110F000.

Updated dictionary.py to correct this encoding step.