Fixed generate_dictionaries.py to no longer point to v1_0_0 of annotations schema (#57)
The step in generate_dicionaries.py to build the binary and map files
for the annotations schema dictionary was pointing to a hard-coded
schema version of v1_0_0, preventing any new annotations added later on
from being included in the dictionary binary. This script now will
attempt to build the dictionary binary using
redfish-payload-annotations-v1.json, which is a clone of the latest
version of the annotation schema.

Also fixed a bug when building a list of annotation schema JSON files
with version less than a given version string could result in
inconsistent ordering of list members, potentially resulting in
indeterministic sequence numbers being assigned to annotations.