Property value type validation for BEJ payload encoding (#67)
Certain BEJ payload encoding scenarios could result in a successful encoding
with invalid data due to missing type checking for the json_data object provided
as a parameter to bej_encode(). For example, attempting to encode a property
that is defined as having a boolean type in the schemas with an assigned value
of "true" will result in a successful encoding, but with the property having a
boolean value of False.

Updated bej_encode_sflv() to ensure that the data type provided in the json_data
object corresponds the the bejFormat assigned to the property in the relevant
schema.